requirements.txt
asgiref==3.8.1
boto3==1.35.68
botocore==1.35.68
dataclasses==0.6
Django==4.2.16
django-cors-headers==4.6.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
djongo==1.3.7
dnspython==2.7.0
drf-yasg==1.21.8
inflection==0.5.1
jmespath==1.0.1
mongoengine==0.29.1
packaging==24.2
PyJWT==2.10.0
pymongo==4.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
s3transfer==0.10.4
six==1.16.0
sqlparse==0.5.2
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3
